let codex apply its taste
