+++ src/cli/app.ts
@@ -1,14 +1,11 @@
-import { Args, Command, Options } from "@effect/cli"
-import { Buffer } from "node:buffer"
-import { createRequire } from "node:module"
-import { readFile } from "node:fs/promises"
-import process from "node:process"
-import { co } from "jazz-tools"
+import { Command, Options } from "@effect/cli"
 import { Effect } from "effect"
-import * as Option from "effect/Option"
+import { createRequire } from "node:module"
 import {
+	acceptDocumentInvite,
 	changeCollaboratorRole,
 	createDocumentInvite,
+	createPersonalDocument,
 	deletePersonalDocument,
 	getDocumentOwner,
 	leavePersonalDocument,
@@ -20,7 +17,6 @@ import {
 	restorePersonalDocument,
 	revokeDocumentInvite,
 } from "@/lib/documents"
-import { acceptDocumentInvite, createPersonalDocument } from "@/lib/documents"
 import {
 	acceptSpaceInvite,
 	changeSpaceCollaboratorRole,
@@ -45,15 +41,28 @@ import {
 	buildSpacePublicLink,
 } from "@/lib/invite-links"
 import { setDocumentTitle } from "@/cli/document-title"
-import { resolveCliConfig } from "@/cli/config"
+import { CliUsageError, PermissionError } from "@/cli/errors"
 import {
-	CliUsageError,
-	NotFoundError,
-	PermissionError,
-	UnexpectedCliError,
-	ValidationError,
-} from "@/cli/errors"
-import { interpretEscapes, parseDocScope, parseScope } from "@/cli/parse"
+	contentFileOption,
+	contentOption,
+	deletedOption,
+	docIdArg,
+	globalOptions,
+	inviteGroupIdOption,
+	linkOption,
+	nameOption,
+	passphraseFileOption,
+	passphraseOption,
+	passphraseStdinOption,
+	roleOption,
+	scopeOption,
+	spaceIdArg,
+	spaceRoleOption,
+	stdinOption,
+	syncOption,
+	titleOption,
+} from "@/cli/options"
+import { parseDocScope } from "@/cli/parse"
 import {
 	createAuthenticatedJazz,
 	generatePassphrase,
@@ -63,110 +72,27 @@ import {
 	logOut,
 	signUpWithPassphrase,
 } from "@/cli/jazz"
-import { printContent, printData, printError } from "@/cli/output"
-import { createSpace, Document, getRandomWriterName } from "@/schema"
+import { printContent } from "@/cli/output"
+import {
+	createSpaceScopedDoc,
+	findDocument,
+	findSpace,
+	getOptionString,
+	inspectInvite,
+	listDocs,
+	loadAccount,
+	maybeSync,
+	readRequiredContentInput,
+	readRequiredSecretInput,
+	readSecretInput,
+	runCommand,
+	summarizeDoc,
+	syncMutation,
+} from "@/cli/runtime"
+import { createSpace, getRandomWriterName } from "@/schema"
 
 export { cli }
 
-type ActiveAccount = Awaited<
-	ReturnType<typeof createAuthenticatedJazz>
->["account"]
-type LoadedAccount = Awaited<ReturnType<typeof loadAccount>>
-
-type GlobalArgs = {
-	json: boolean
-	verbose: boolean
-	quiet: boolean
-	server: Option.Option<string>
-	syncPeer: Option.Option<string>
-	timeout: Option.Option<number>
-	home: Option.Option<string>
-}
-
-type LoadedCliDocument = co.loaded<typeof Document, { content: true }>
-
-let jsonOption = Options.boolean("json").pipe(
-	Options.withDescription("Print machine-readable JSON output."),
-)
-let verboseOption = Options.boolean("verbose").pipe(
-	Options.withDescription("Include extra runtime context in output."),
-)
-let quietOption = Options.boolean("quiet").pipe(
-	Options.withDescription("Suppress output. Exit code only."),
-)
-let serverOption = Options.optional(Options.text("server")).pipe(
-	Options.withDescription(
-		"Alkalye deployment URL. Used to discover CLI config and sync peer.",
-	),
-)
-let syncPeerOption = Options.optional(Options.text("sync-peer")).pipe(
-	Options.withDescription(
-		"Jazz sync peer URL. Defaults to ALKALYE_SYNC_PEER or PUBLIC_JAZZ_SYNC_SERVER.",
-	),
-)
-let timeoutOption = Options.optional(Options.integer("timeout")).pipe(
-	Options.withDescription("Sync timeout in milliseconds."),
-)
-let homeOption = Options.optional(Options.text("home")).pipe(
-	Options.withDescription(
-		"CLI state directory. Defaults to ALKALYE_CLI_HOME or ~/.alkalye/cli.",
-	),
-)
-
-let globalOptions = {
-	json: jsonOption,
-	verbose: verboseOption,
-	quiet: quietOption,
-	server: serverOption,
-	syncPeer: syncPeerOption,
-	timeout: timeoutOption,
-	home: homeOption,
-}
-
-let nameOption = Options.text("name").pipe(
-	Options.withDescription("Name to store."),
-)
-let titleOption = Options.text("title").pipe(
-	Options.withDescription("Document title."),
-)
-let roleOption = Options.choice("role", ["writer", "reader"]).pipe(
-	Options.withDescription("Invite role."),
-)
-let spaceRoleOption = Options.choice("role", [
-	"admin",
-	"manager",
-	"writer",
-	"reader",
-]).pipe(Options.withDescription("Space collaborator role."))
-let linkOption = Options.text("link").pipe(
-	Options.withDescription("Invite link to inspect or accept."),
-)
-let inviteGroupIdOption = Options.text("invite-group-id").pipe(
-	Options.withDescription("Invite group ID."),
-)
-let scopeOption = Options.optional(Options.text("scope")).pipe(
-	Options.withDescription("Document scope: personal, all, or space:<id>."),
-)
-let deletedOption = Options.boolean("deleted").pipe(
-	Options.withDescription("Include deleted documents instead of active ones."),
-)
-let syncOption = Options.boolean("sync").pipe(
-	Options.withDescription("Wait for remote sync before exiting."),
-)
-let passphraseOption = Options.optional(Options.text("passphrase"))
-let passphraseFileOption = Options.optional(Options.file("passphrase-file"))
-let passphraseStdinOption = Options.boolean("passphrase-stdin").pipe(
-	Options.withDescription("Read passphrase from stdin."),
-)
-let contentOption = Options.optional(Options.text("content"))
-let contentFileOption = Options.optional(Options.file("content-file"))
-let stdinOption = Options.boolean("stdin")
-let docIdArg = Args.text({ name: "doc-id" }).pipe(
-	Args.withDescription("Document ID."),
-)
-let spaceIdArg = Args.text({ name: "space-id" }).pipe(
-	Args.withDescription("Space ID."),
-)
 let packageJson = createRequire(import.meta.url)("../../package.json") as {
 	version?: string
 }
@@ -335,7 +261,16 @@ let accountCommand = Command.make("account").pipe(
 	Command.withSubcommands([accountShow, accountRename, accountSync]),
 )
 
-let syncFlush = Command.make("flush", globalOptions, accountSync.handler)
+let syncFlush = Command.make("flush", globalOptions, args =>
+	runCommand("sync.flush", args, async config => {
+		let jazz = await createAuthenticatedJazz(config)
+		await jazz.account.$jazz.waitForAllCoValuesSync({
+			timeout: config.timeoutMs,
+		})
+		await jazz.done()
+		return { synced: true, syncPeer: config.syncPeer }
+	}),
+)
 
 let syncStatus = Command.make("status", globalOptions, args =>
 	runCommand("sync.status", args, async config => {
@@ -1198,334 +1133,3 @@ let cli = Command.run(root, {
 	name: "alkalye",
 	version: packageJson.version ?? "0.0.0",
 })
-
-function runCommand<A extends GlobalArgs>(
-	command: string,
-	args: A,
-	handler: (
-		config: Awaited<ReturnType<typeof resolveCliConfig>>,
-	) => Promise<unknown>,
-) {
-	return Effect.tryPromise({
-		try: async () => {
-			try {
-				let config = await resolveFlags(args)
-				let data = await handler(config)
-				if (config.quiet || data === undefined) return
-				await Effect.runPromise(
-					printData({
-						json: config.json,
-						command,
-						data,
-						meta: config.verbose
-							? {
-									serverUrl: config.serverUrl,
-									syncPeer: config.syncPeer,
-									timeoutMs: config.timeoutMs,
-									homeDir: config.homeDir,
-								}
-							: undefined,
-					}),
-				)
-			} catch (error) {
-				printError({
-					json: args.json,
-					command,
-					error: {
-						type: getErrorType(error),
-						message: getErrorMessage(error),
-					},
-				})
-				throw error
-			}
-		},
-		catch: error =>
-			new UnexpectedCliError({
-				message: getErrorMessage(error),
-			}),
-	})
-}
-
-function resolveFlags(args: GlobalArgs) {
-	return resolveCliConfig({
-		json: args.json,
-		verbose: args.verbose,
-		quiet: args.quiet,
-		server: getOptionString(args.server),
-		syncPeer: getOptionString(args.syncPeer),
-		timeout: getOptionNumber(args.timeout),
-		home: getOptionString(args.home),
-	})
-}
-
-async function loadAccount(account: ActiveAccount, timeoutMs: number = 10_000) {
-	await bestEffortRefresh(account, timeoutMs)
-	return account.$jazz.ensureLoaded({
-		resolve: {
-			profile: true,
-			root: {
-				documents: { $each: { content: true } },
-				inactiveDocuments: { $each: { content: true } },
-				spaces: { $each: { documents: { $each: { content: true } } } },
-			},
-		},
-	})
-}
-
-async function bestEffortRefresh(account: ActiveAccount, timeoutMs: number) {
-	try {
-		await account.$jazz.waitForAllCoValuesSync({ timeout: timeoutMs })
-	} catch {
-		return
-	}
-}
-
-async function listDocs(
-	account: LoadedAccount,
-	scopeValue: string | undefined,
-	deleted: boolean,
-) {
-	let scope = parseScope(scopeValue)
-	let entries = collectDocsForScope(account, scope)
-		.flatMap(entry =>
-			entry.doc ? [{ doc: entry.doc, spaceId: entry.spaceId }] : [],
-		)
-		.filter(entry =>
-			deleted ? Boolean(entry.doc.deletedAt) : !entry.doc.deletedAt,
-		)
-	let summaries = entries.map(entry => summarizeDoc(entry.doc, entry.spaceId))
-	return summaries.sort((left, right) =>
-		right.updatedAt.localeCompare(left.updatedAt),
-	)
-}
-
-function collectDocsForScope(
-	account: LoadedAccount,
-	scope: ReturnType<typeof parseScope>,
-) {
-	if (scope.kind === "space") {
-		let space = findSpace(account, scope.spaceId)
-		return space.documents.map(doc => ({ doc, spaceId: space.$jazz.id }))
-	}
-	let personal = [
-		...account.root.documents,
-		...(account.root.inactiveDocuments ?? []),
-	].map(doc => ({ doc, spaceId: null as string | null }))
-	if (scope.kind === "personal") return personal
-	let spaceDocs = (account.root.spaces ?? []).flatMap(space =>
-		space.documents.map(doc => ({
-			doc,
-			spaceId: space.$jazz.id as string | null,
-		})),
-	)
-	return [...personal, ...spaceDocs]
-}
-
-async function findDocument(account: LoadedAccount, docId: string) {
-	for (let doc of account.root.documents) {
-		if (doc?.$jazz.id === docId)
-			return { doc: await ensureDocLoaded(doc), space: null }
-	}
-	for (let doc of account.root.inactiveDocuments ?? []) {
-		if (doc?.$jazz.id === docId)
-			return { doc: await ensureDocLoaded(doc), space: null }
-	}
-	for (let space of account.root.spaces ?? []) {
-		for (let doc of space.documents) {
-			if (doc?.$jazz.id === docId) {
-				return { doc: await ensureDocLoaded(doc), space }
-			}
-		}
-	}
-	throw new NotFoundError({ message: `Document not found: ${docId}` })
-}
-
-function findSpace(account: LoadedAccount, spaceId: string) {
-	let space = account.root.spaces?.find(item => item?.$jazz.id === spaceId)
-	if (!space)
-		throw new NotFoundError({ message: `Space not found: ${spaceId}` })
-	return space
-}
-
-async function createSpaceScopedDoc(
-	account: LoadedAccount,
-	spaceId: string,
-	content: string,
-) {
-	let space = findSpace(account, spaceId)
-	let doc = await createPersonalDocument(account, content)
-	await moveDocumentToSpace({
-		doc,
-		destination: { id: space.$jazz.id, name: space.name },
-		me: account,
-	})
-	return doc
-}
-
-function summarizeDoc(
-	doc: {
-		$jazz: { id: string }
-		content?: { toString(): string }
-		createdAt: Date
-		updatedAt: Date
-		deletedAt?: Date
-		spaceId?: string
-	},
-	spaceId?: string | null,
-) {
-	return {
-		docId: doc.$jazz.id,
-		title: getDocumentTitle(doc),
-		spaceId: spaceId ?? doc.spaceId ?? null,
-		createdAt: doc.createdAt.toISOString(),
-		updatedAt: doc.updatedAt.toISOString(),
-		deletedAt: doc.deletedAt?.toISOString() ?? null,
-	}
-}
-
-async function maybeSync(
-	account: ActiveAccount,
-	enabled: boolean,
-	timeoutMs: number,
-) {
-	if (!enabled) return
-	await account.$jazz.waitForAllCoValuesSync({ timeout: timeoutMs })
-}
-
-async function syncMutation(account: ActiveAccount, timeoutMs: number) {
-	await account.$jazz.waitForAllCoValuesSync({ timeout: timeoutMs })
-}
-
-async function readRequiredContentInput(
-	content: Option.Option<string>,
-	contentFile: Option.Option<string>,
-	stdin: boolean,
-) {
-	let next = await readContentInput(content, contentFile, stdin)
-	if (next === undefined) {
-		throw new CliUsageError({
-			message: "Provide one of --content, --content-file, --stdin",
-		})
-	}
-	return next
-}
-
-async function readContentInput(
-	content: Option.Option<string>,
-	contentFile: Option.Option<string>,
-	stdin: boolean,
-): Promise<string | undefined> {
-	let sources = [
-		Option.isSome(content),
-		Option.isSome(contentFile),
-		stdin,
-	].filter(Boolean)
-	if (sources.length > 1) {
-		throw new CliUsageError({ message: "Provide exactly one content source" })
-	}
-	if (Option.isSome(content)) return interpretEscapes(content.value)
-	if (Option.isSome(contentFile))
-		return await readFile(contentFile.value, "utf8")
-	if (stdin) return await readStdin()
-	return undefined
-}
-
-async function readRequiredSecretInput(
-	passphrase: Option.Option<string>,
-	passphraseFile: Option.Option<string>,
-	passphraseStdin: boolean,
-) {
-	let next = await readSecretInput(passphrase, passphraseFile, passphraseStdin)
-	if (!next) {
-		throw new CliUsageError({
-			message:
-				"Provide one of --passphrase, --passphrase-file, --passphrase-stdin",
-		})
-	}
-	return next
-}
-
-async function readSecretInput(
-	passphrase: Option.Option<string>,
-	passphraseFile: Option.Option<string>,
-	passphraseStdin: boolean,
-): Promise<string | undefined> {
-	let sources = [
-		Option.isSome(passphrase),
-		Option.isSome(passphraseFile),
-		passphraseStdin,
-	].filter(Boolean)
-	if (sources.length > 1) {
-		throw new CliUsageError({
-			message: "Provide exactly one passphrase source",
-		})
-	}
-	if (Option.isSome(passphrase)) return passphrase.value.trim()
-	if (Option.isSome(passphraseFile)) {
-		return (await readFile(passphraseFile.value, "utf8")).trim()
-	}
-	if (passphraseStdin) return (await readStdin()).trim()
-	return undefined
-}
-
-function inspectInvite(link: string) {
-	try {
-		let invite = parseInviteLink(link)
-		return { kind: "doc", ...invite }
-	} catch (error) {
-		void error
-	}
-	try {
-		let invite = parseSpaceInviteLink(link)
-		return { kind: "space", ...invite }
-	} catch (error) {
-		void error
-	}
-	throw new ValidationError({ message: "Invalid invite link" })
-}
-
-function getOptionString(value: Option.Option<string> | undefined) {
-	return value && Option.isSome(value) ? value.value : undefined
-}
-
-function getOptionNumber(value: Option.Option<number> | undefined) {
-	return value && Option.isSome(value) ? value.value : undefined
-}
-
-function getErrorType(error: unknown): string {
-	return typeof error === "object" &&
-		error !== null &&
-		"_tag" in error &&
-		typeof error._tag === "string"
-		? error._tag
-		: "Error"
-}
-
-function getErrorMessage(error: unknown): string {
-	return error instanceof Error ? error.message : String(error)
-}
-
-async function readStdin(): Promise<string> {
-	let chunks: Buffer[] = []
-	for await (let chunk of process.stdin) {
-		chunks.push(typeof chunk === "string" ? Buffer.from(chunk) : chunk)
-	}
-	return Buffer.concat(chunks).toString("utf8")
-}
-
-async function ensureDocLoaded(
-	doc:
-		| LoadedCliDocument
-		| {
-				$jazz: {
-					ensureLoaded(args: {
-						resolve: { content: true }
-					}): Promise<LoadedCliDocument>
-				}
-		  },
-): Promise<LoadedCliDocument> {
-	if ("content" in doc) {
-		return doc
-	}
-	return doc.$jazz.ensureLoaded({ resolve: { content: true } })
-}
